requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
sentence-transformers>=3.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
huggingface-hub>=0.20.0